refactor: substitui métodos deprecated do Filament
