+++ app/Filament/Resources/RifaResource/Pages/EditRifa.php
@@ -3,7 +3,7 @@
 namespace App\Filament\Resources\RifaResource\Pages;
 
 use App\Filament\Resources\RifaResource;
-use Filament\Pages\Actions;
+use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
 
 class EditRifa extends EditRecord
